+++ sparse_strips/vello_example_scenes/src/lib.rs
@@ -21,9 +21,10 @@ pub mod svg;
 pub mod text;
 
 use glifo::GlyphRunBackend;
+pub use vello_common::blurred_rounded_rect::CornerRadii;
 use vello_common::filter_effects::Filter;
 use vello_common::kurbo::Affine;
-pub use vello_common::kurbo::{BezPath, Rect, RoundedRectRadii, Shape, Stroke};
+pub use vello_common::kurbo::{BezPath, Rect, Shape, Stroke};
 pub use vello_common::mask::Mask;
 use vello_common::paint::ImageSource;
 pub use vello_common::paint::{Paint, PaintType};
@@ -81,7 +82,7 @@ pub trait RenderingContext: Sized {
     /// Fill a rectangle with the current paint.
     fn fill_rect(&mut self, rect: &Rect);
     /// Fill a blurred rounded rectangle with the current solid paint.
-    fn fill_blurred_rounded_rect(&mut self, rect: &Rect, radii: RoundedRectRadii, std_dev: f32);
+    fn fill_blurred_rounded_rect(&mut self, rect: &Rect, radii: CornerRadii, std_dev: f32);
     /// Create a glyph run builder for text rendering.
     fn glyph_run<'a>(
         &'a mut self,
@@ -172,7 +173,7 @@ impl RenderingContext for RenderContext {
         self.fill_rect(rect);
     }
 
-    fn fill_blurred_rounded_rect(&mut self, rect: &Rect, radii: RoundedRectRadii, std_dev: f32) {
+    fn fill_blurred_rounded_rect(&mut self, rect: &Rect, radii: CornerRadii, std_dev: f32) {
         self.fill_blurred_rounded_rect(rect, radii, std_dev, false);
     }
 
@@ -277,7 +278,7 @@ impl RenderingContext for Scene {
         self.fill_rect(rect);
     }
 
-    fn fill_blurred_rounded_rect(&mut self, rect: &Rect, radii: RoundedRectRadii, std_dev: f32) {
+    fn fill_blurred_rounded_rect(&mut self, rect: &Rect, radii: CornerRadii, std_dev: f32) {
         self.fill_blurred_rounded_rect(rect, radii, std_dev, false);
     }